requirements.txt
pymongo==4.6.1
transformers>=4.30.0
torch>=1.13.0
torchvision>=0.14.0
pillow>=9.0.0
sentence-transformers>=2.2.0
huggingface_hub>=0.15.1
numpy>=1.21.0
pandas>=1.3.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0
requests>=2.28.0